+++ integrations/sentry/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 <!-- towncrier release notes start -->
 
+## 0.2.0 (2025-08-25)
+
+
+- Added rate limiter with using `RetryConfig`
+
+
 ## 0.1.198 (2025-08-28)